LayoutTests/platform/ios-simulator/editing/execCommand/format-block-table-expected.txt


Formatting:
| "
"
| <table>
|   <tbody>
|     <tr>
|       <td>
|         "hello"
|       <td>
|         "world"
| "
"

by p yields:
| "
"
| <table>
|   <tbody>
|     <tr>
|       <td>
|         "hello"
|       <td>
|         "world"
| "
"

Formatting:
| "
"
| <table>
|   <tbody>
|     <tr>
|       <td>
|         "<#selection-anchor>hello<#selection-focus>"
|       <td>
|         "world"
| "
"

by blockquote yields:
| "
"
| <table>
|   <tbody>
|     <tr>
|       <td>
|         <blockquote>
|           "hello"
|       <td>
|         "world"
| "
"

Formatting:
| "
"
| <table>
|   <tbody>
|     <tr>
|       <td>
|         <p>
|           "hello"
|         <div>
|           "world"
|       <td>
|         "WebKit"
| "
"

by h3 yields:
| "
"
| <table>
|   <tbody>
|     <tr>
|       <td>
|         <p>
|           "hello"
|         <div>
|           "world"
|       <td>
|         "WebKit"
| "
"

Formatting:
| "
"
| <table>
|   <tbody>
|     <tr>
|       <td>
|         <ul>
|           <li>
|             "hello"
|           <li>
|             "world"
|       <td>
|         "WebKit"
| "
"

by address yields:
| "
"
| <table>
|   <tbody>
|     <tr>
|       <td>
|         <ul>
|           <li>
|             "hello"
|           <li>
|             "world"
|       <td>
|         "WebKit"
| "
"
